requirements.txt
pyramid
pyramid_jinja2
pyramid_debugtoolbar
waitress
sqlalchemy
celery
zope.sqlalchemy
mock
nose